Merge pull request #528 from xoox:calib-release-object
Add new test data for extended camera calibration method (#528)

* Reorder object points and image points

Now the feature points are in the order as the same set of the object
points, which makes the test data usable for testing calibration with
CALIB_RELEASE_OBJECT flag.

* Add one line to store calibration flags

opencv/modules/calib3d/test/test_cameracalibration.cpp must be sync with
this test data file.

* Add new test data for camera calibration

The new added calib2.dat is used for regression test of camera
calibration with CALIB_RELEASE_OBJECT flag.

* Add a line of data for the index of fixed point

The new proposed calibration flag CALIB_RELEASE_OBJECT is removed. And
the method selection is based on the range of the index of fixed point.
For minus values, standard calibration method will be chosen. Values in
a rational range will make the object-releasing calibration method
selected.