put-gput group norm tests into conformance opencv_extra
Attention: The data provided here differ from the original data found in onnx/onnx/backend/test/data/node. In response to the issue outlined in [ONNX Issue #5466](https://github.com/onnx/onnx/issues/5466), I have generated new input and output datasets for the OpenCV test data. It's important to note that the tensor shape of the scale and bias are based on the C (channels) parameter.